[gfx] Separate components into different structures
In the previous commit, the initial code for creating and rendering into
windows was provided, but now this code is separated into logical,
separate components that reflect common naming schemes and structures of
DirectX, Vulkan, and WebGPU, as they are resposible for the underlying
data that they hold. E.g. the `RenderPass` initializes, uses, and
finishes a render pass.