[layout] Sketch out basic `Box` structure
CSS uses a (layout) boxes tree, which are generated based on the
DOM tree. This commit introduces theses box types, which are generated
based on the `display` property of box nodes.

I've documented the applicable CSS specifications, but most of the code
is based on CSS 2. CSS 2 is actually still primarily used for the Visual
Formatting Model and the Box Model, post-CSS 2 (colloquially known as
CSS 3) only midly increments upon these.