[dom] Use [`Tendril`] instead of `String` for text nodes
The data of text nodes (e.g. the actual `String`s) is used for layout to
represent the text that is laid out.

The problem is that these layout
nodes are often split into multiple line boxes, for example when a
paragraph is split up into multiple lines.

For this, there must be multiple line boxes that each contain the
substring for that portion of the line, but that would require either
cloning the `String` (which results in a lot of small strings), or
taking a `&str` slice of the `String`. The latter is impossible because
of the borrow checker, since we cannot easily borrow the `String`s in a
separate tree structure.

This is where [`Tendril`] comes into play, which uses reference counted
strings with internal slicing, which is like the [`FlyString`] from
SerenityOS.

[`FlyString`]: https://github.com/SerenityOS/serenity/blob/master/AK/FlyString.h
[`Tendril`]: https://crates.io/crates/tendril