[main] Default log level to `warn` on debug builds
Using `env_logger::init()`, the default log level is only `error`. This
can be modified by the `RUST_LOG` environment variable, but it is good
practice for debug builds to display warnings, errors and infos.

This can be modified to silence the logs or extend the logs, whether in
debug builds or release builds, since `RUST_LOG` always takes
precedence.