[gfx] Fix incorrect position and size of painting a colored rect
I actually believed to have fixed this in a previous commit, but this
time, I am sure it is fixed ^_^.

The position and size were incorrect due to an incorrect matrix
transform calculation, which seemed to be correct in a few test cases,
but didn't scale up to individual boxes. It now uses an orthographical
projection, combined with a view matrix, to correctly compute the size
and position for the shader.

The vertices of the colored_vertex.rs are now also correct, whereas they
previously would be too big.