[page+layout] Don't regenerate layout tree on resize
We can just rerun the formatting contexts, instead of doing a whole new
calculation.

We do however need to recalculate the style, because `100vw` isn't the
same on a resize, for example.

Also, when we support `@media` queries, we might be required to
regenerate the layout tree since the media query might specify a pseudo,
other display value, etc.