[gfx] Prevent triple resize event when the window opens
This is a rather weird bug, since I can't find anything related to this
online - neither the GitHub of winit, nor the Win32 forums.

The fix is to ignore all resize events (on Windows i.e. WS_SIZE) during
the first 250 milliseconds from when the Window is opened. I assume this
won't break anything, since users are (probably) unlikely to modify the
size of the window in such a short time. If this breaks something, a
better solutions is required.