[gfx] Render textured rects using the actual given rect
Previously, this API would ignore the rect and just fill the whole
viewport. However, since this function is used for filling the window
surface with the page surface, it is better to actually use the rect.
When resizing, the window can use the texture view from the last paint
phase before the page canvas is repainted, and avoid stretching out that
texture view.