[compositor] Keep bottom tiles in cache when scrolling up
The tiles are kept in the cache by the maximum y coordinate of the
viewport. This ensure that the topmost tiles won't be deleted when the
viewport is moved down by scrolling.

The problem was however that the max y of the viewport would decrease
when scrolling up, meaning that a `Vec::resize_with` would shrink.

We now ensure that the composited tiles won't be invalidated when
scrolling up, by checking whether or not the new capacity is greater
than the current capacity.