[compositor] Delay waiting on submit until next composite cycle
Before this commit, we would wait on the GPU to finish its work with
compositing the tiles before accepting new ones for a next cycle. This
was reducing the parallelization of the CPU and GPU.

Now, we submit the command buffers but only wait for them to finish (and
the call the upload_image_callback) when we want to recomposite again in
a new cycle.