[font] Improve glyph rendering by eliminating double AA
This patch fixes a longstanding annoyance of mine, the fact that the
text rendered by DirectWrite was very pixelated. I consider DWrite to be
pretty good software, but my results just weren't it.

I've tried for countless hours to debug this issue, and I've finally
found the problem: due to kerning and other glyph offsetting, the
position of the glyphs would become subpixel.

Meaning they would be rendered at (99.58514, 99.0085). This resulted in
the GPU sampling a non-linear glyph texel, and aliasing it to scale to
integer pixels on the render target. This was very unnecessary, since
the glyphs are already rendered to the render target using the same
exact bitmap dimensions of the rasterized glyph texture.

By rounding the coordinates of the textured rect, we avoid the above
problems, which results in much more readible text (especially at small
scales).