[page] Reduce maximum delay between paints
Initially, this value was around 250 ms, and now it is 30 ms. I chose
this big value, to ensure between the page load and the load of
resources, enough time was passed to reduce FUOCs on each load.

The value now, 30 ms, is chosen because when no resources are being
loaded/awaited, repainting after a scroll event can take around 100 ms!
Regardless of efforts of improving paint/composite speed, this would
limit the windows FPS to a measily 13 fps.

While 30 ms will not result in a smooth 120+ frame rate, this does
ensure at least 30 fps is reached. In the future there should definitely
be work put in to improve the frame rate.

TODO: Allow layout, paint, and composition to occur concurrently, and,
      if possible, allow multiple paints to be in flight at the same
      time. - currently this is impossible due to the nature of the
      page system.