[font] Fix crash when using FreeType
DirectWrite resources can be used across threads, but unfortunately
FreeType can't. This bugfix removes the parallel preparation for glyph
rasterization when using FreeType, but it might be possible in the
future to initialize a thread pool and load the FT_Face ourselves by
sharing the buffer and parallelizing the rasterization thereby.