[gfx] Perf: share MaterialRenderers across Tiles
Previously, each Tile would create its own CanvasRenderingContext, thus
an Artwork, thus a TextureMaterialRenderer and a ColorMaterialRenderer.

Although this seemed necessary at first - namely because each
MaterialRenderer has a single Uniform buffer - but this was not
necessary.

The majority of time rendering the page was actually taken up by the
creation of these structures, so this is a huge performance boost! :)