[layout+dom+page] Share images with the same source
On sites like Wikipedia, it is common to use many images with the same
URL (such as a link icon). This change introduces the ImageProvider,
which retrieves the image from a shared cache to see if ImageData is
already associated with the given URL.

This improves memory usage, since we don't have to store these
seperately on the GPU, but also network and CPU usage since we only use
one request for all these image instances.

In the future, it is also made easier to implement client-side caching
for images by using this facade. We can then just store those ImageData
by some alias and when those images would get requested from the
network, we just load them from disk.