[layout] Avoid crashing when a Text node doesn't have a parent
Finding out the language for a given Text node, as is necessary for
correct `text-transform`ation, is done by looking up the `lang`
attribute of the Text's parent element.

This would crash when adding an anonymous replaced text node for the
contents of a `<input type="text">`, because that wouldn't be associated
with a real dom Text node, but with an emulated one that doesn't messes
with the DOM at all.