[font] Remove temporary fork of font-kit and dwrote
These were initially forked to better support certain features (mostly
emoji support), but given the complexity of actually using a colored
glyph run, it is better to render these using an SVG-like renderer
instead of using the DirectWrite/2D-specific calls. The latter is
because we would need to render to a Direct2D-surface (but that is
abstracted away using wgpu, and can also be OpenGL or Vulkan instead).